Refactor BackButton file structure
