+++ src/components/BackButton/index.js
@@ -1,3 +1,17 @@
+import { connect } from 'react-redux';
+import { injectIntl } from 'react-intl';
 import BackButton from './BackButton';
 
-export default BackButton;
+// Listen to redux state
+const mapStateToProps = (state) => {
+  const { breadcrumb, navigator } = state;
+  return {
+    breadcrumb,
+    navigator,
+  };
+};
+
+export default injectIntl(connect(
+  mapStateToProps,
+  null,
+)(BackButton));
